"Update views and permissions for cadastro, erickgym and treinos apps; remove treinos/urls copy.py"
This diff includes changes to views and permissions in the cadastro, erickgym, and treinos apps. In the cadastro app, the `permissions.IsAuthenticated` permission has been changed to `IsAdminOrReadOnly` for the `ListCreateAlunosAPIView` and `ListCreateProfessor` classes. The `permissions.IsAuthenticated` permission has been changed to `IsAdmin | IsOwner` for the `RetrieveUpdateDestryAlunoAPIView` and `RetrieveUpdateDestroyProfessor` classes.

In the erickgym app, a new `IsOwner` permission has been added, and the `IsOwnerOrReadOnly` and `IsAdmin` permissions have been updated. The `UserSerializer` has been renamed to `CommomUserSerializer`.

In the treinos app, the `permissions.IsAuthenticated` permission has been changed to `permissions.IsAuthenticatedOrReadOnly` for the `ListCreateExercicioAPIView` class. The `permissions.IsAuthenticated` permission has been changed to `IsOwnerOrReadOnly | IsAdminOrReadOnly` for the `DetailUpdateDeleteExercicio` class. The `treinos/urls copy.py` file has been deleted.

Additionally, the imports in the `erickgym/views.py` file have been updated to include `BasicAuthentication`. The `login` and `singup` views now use `BasicAuthentication` instead of `SessionAuthentication`.